(submitted v3) contrib: init: sysvinit-debian: rename $DOCKERD to $DAEMON
Make it a bit more aligned to debian init script standards. The idea behind is
is all init scripts should look more or less the same, with minimal difference.
(eg. for easier reviews, linting, etc) The name of the actual daemon binary is
put into the name DAEMON.

Therefore fix it up to be a little bit aligned better to debian standards.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>