(submitted) libnetwork: iptables: make systemd/firewalld specific hacks optional
When running on a pure Linux machine (as opposed to Lennartix), we don't
need any systemd specific code - in that case it's just unnecessary
complexity and a lot of code (with dependencies around 10kLoC) that
needs to be maintained and audited.

Therefore make it all optional by setting 'no_systemd' build tag.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>